+++ packages/fuels/tests/harness.rs
@@ -926,7 +926,7 @@ async fn test_provider_launch_and_connect() -> Result<(), Error> {
         DEFAULT_NUM_COINS,
         DEFAULT_COIN_AMOUNT,
     );
-    let (launched_provider, address) = setup_test_provider(coins, None, None).await;
+    let (launched_provider, address) = setup_test_provider(coins, None).await;
     let connected_provider = Provider::connect(address).await?;
 
     wallet.set_provider(connected_provider);
@@ -1511,7 +1511,7 @@ async fn test_wallet_balance_api() -> Result<(), ProviderError> {
         amount_per_coin,
     );
 
-    let (provider, _) = setup_test_provider(coins.clone(), None, None).await;
+    let (provider, _) = setup_test_provider(coins.clone(), None).await;
     wallet.set_provider(provider);
     for (_utxo_id, coin) in coins {
         let balance = wallet.get_asset_balance(&coin.asset_id).await;
@@ -1538,7 +1538,7 @@ async fn test_wallet_balance_api() -> Result<(), ProviderError> {
     );
     assert_eq!(coins.len() as u64, number_of_assets * coins_per_asset);
     assert_eq!(asset_ids.len() as u64, number_of_assets);
-    let (provider, _) = setup_test_provider(coins.clone(), None, None).await;
+    let (provider, _) = setup_test_provider(coins.clone(), None).await;
     wallet.set_provider(provider);
     let balances = wallet.get_balances().await?;
     assert_eq!(balances.len() as u64, number_of_assets);
